Move phone location toggle to settings
